Optimizations
Cached CFGS by zone/area id so we're not calling the CFG every single kill. That was like ~90% of the load. Added a cache flush to the gm command for on-the-fly test editing. Otherwise that feature would no longer work.

Item scan becomes the biggest performance suck now but I dont know how to make that better yet.

Fixed an issue with webhook, multiple players dying on same frame would not register to Discord. It works as it should now.

Removed a worldobject find from a for loop, that was dumb to have done.

Dumpped LootStore & ChestGold tables shortly after chests despawn so memory doesnt get eaten up over time.

fixed a couple dividers aesthetically